# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "오뚜기 파워세일즈 API",
    "description": "B2B 영업사원 실적 조회 및 목표 관리 시스템 API",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:49275",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
  "paths": {
    "/api/v1/mobile/product-expiration/{seq}": {
      "put": {
        "tags": [
          "product-expiration-controller"
        ],
        "operationId": "updateProductExpiration",
        "parameters": [
          {
            "name": "seq",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductExpirationUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProductExpirationItemResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "product-expiration-controller"
        ],
        "operationId": "deleteProductExpiration",
        "parameters": [
          {
            "name": "seq",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mobile/safety-check/submit": {
      "post": {
        "tags": [
          "safety-check-controller"
        ],
        "operationId": "submitSafetyCheck",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SafetyCheckSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSafetyCheckSubmitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mobile/product-expiration": {
      "get": {
        "tags": [
          "product-expiration-controller"
        ],
        "operationId": "getProductExpirationList",
        "parameters": [
          {
            "name": "accountCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListProductExpirationItemResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "product-expiration-controller"
        ],
        "operationId": "createProductExpiration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductExpirationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProductExpirationItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mobile/product-expiration/batch-delete": {
      "post": {
        "tags": [
          "product-expiration-controller"
        ],
        "operationId": "deleteProductExpirationBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductExpirationBatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProductExpirationBatchDeleteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mobile/attendance": {
      "post": {
        "tags": [
          "attendance-controller"
        ],
        "operationId": "registerCommute",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCommuteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mobile/safety-check/status": {
      "get": {
        "tags": [
          "safety-check-controller"
        ],
        "operationId": "getStatus",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSafetyCheckStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mobile/products/search": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "operationId": "searchProducts",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "text"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageProductDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mobile/attendance/accounts": {
      "get": {
        "tags": [
          "attendance-controller"
        ],
        "operationId": "getAccountList",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAccountListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/safety-check/status": {
      "get": {
        "tags": [
          "admin-safety-check-controller"
        ],
        "operationId": "getStatus_1",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSafetyCheckStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProductExpirationUpdateRequest": {
        "type": "object",
        "properties": {
          "expirationDate": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "alarmDate": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "alarmDate",
          "expirationDate"
        ]
      },
      "ApiResponseProductExpirationItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ProductExpirationItemResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ProductExpirationItemResponse": {
        "type": "object",
        "properties": {
          "seq": {
            "type": "integer",
            "format": "int32"
          },
          "productCode": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "accountCode": {
            "type": "string"
          },
          "accountName": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string"
          },
          "alarmDate": {
            "type": "string"
          },
          "dDay": {
            "type": "integer",
            "format": "int32",
            "writeOnly": true
          },
          "description": {
            "type": "string"
          },
          "isExpired": {
            "type": "boolean"
          },
          "dday": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "accountCode",
          "accountName",
          "alarmDate",
          "dday",
          "expirationDate",
          "isExpired",
          "productCode",
          "productName",
          "seq"
        ]
      },
...
      "ApiResponseObject": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
      "EquipmentAnswer": {
        "type": "object",
        "properties": {
          "seqNum": {
            "type": "integer",
            "format": "int32"
          },
          "answer": {
            "type": "string"
          }
        },
        "required": [
          "answer",
          "seqNum"
        ]
      },
      "SafetyCheckSubmitRequest": {
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "completeTime": {
            "type": "string",
            "format": "date-time"
          },
          "equipments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EquipmentAnswer"
            }
          },
          "precautions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "completeTime",
          "equipments",
          "startTime"
        ]
      },
      "ApiResponseSafetyCheckSubmitResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/SafetyCheckSubmitResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "SafetyCheckSubmitResponse": {
        "type": "object",
        "properties": {
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "safetyCheckCompleted": {
            "type": "boolean"
          }
        },
        "required": [
          "safetyCheckCompleted",
          "submittedAt"
        ]
      },
      "ProductExpirationCreateRequest": {
        "type": "object",
        "properties": {
          "accountCode": {
            "type": "string"
          },
          "accountName": {
            "type": "string"
          },
          "productCode": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "alarmDate": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "accountCode",
          "accountName",
          "alarmDate",
          "expirationDate",
          "productCode",
          "productName"
        ]
      },
      "ProductExpirationBatchDeleteRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "ApiResponseProductExpirationBatchDeleteResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ProductExpirationBatchDeleteResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "ProductExpirationBatchDeleteResponse": {
        "type": "object",
        "properties": {
          "deletedCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "deletedCount"
        ]
      },
...
      "CommuteRequest": {
        "type": "object",
        "properties": {
          "scheduleId": {
            "type": "integer",
            "format": "int64"
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "maximum": 90,
            "minimum": -90
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "maximum": 180,
            "minimum": -180
          },
          "workType": {
            "type": "string"
          }
        },
        "required": [
          "latitude",
          "longitude",
          "scheduleId"
        ]
      },
      "ApiResponseCommuteResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/CommuteResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "CommuteResponse": {
        "type": "object",
        "properties": {
          "scheduleId": {
            "type": "integer",
            "format": "int64"
          },
          "accountName": {
            "type": "string"
          },
          "workType": {
            "type": "string"
          },
          "distanceKm": {
            "type": "number",
            "format": "double"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "registeredCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "accountName",
          "distanceKm",
          "registeredCount",
          "scheduleId",
          "totalCount"
        ]
      },
...
      "ApiResponseSafetyCheckStatusResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/SafetyCheckStatusResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
      "SafetyCheckStatusResponse": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "submittedCount": {
            "type": "integer",
            "format": "int32"
          },
          "notSubmittedCount": {
            "type": "integer",
            "format": "int32"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberStatus"
            }
          }
        },
        "required": [
          "date",
          "members",
          "notSubmittedCount",
          "submittedCount",
          "totalCount"
        ]
      },
...
      "ApiResponsePageProductDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/PageProductDto"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "PageProductDto": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponseListProductExpirationItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductExpirationItemResponse"
            }
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
      "AccountListResponse": {
        "type": "object",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountInfo"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "registeredCount": {
            "type": "integer",
            "format": "int32"
          },
          "currentDate": {
            "type": "string"
          }
        },
        "required": [
          "accounts",
          "currentDate",
          "registeredCount",
          "totalCount"
        ]
      },
      "ApiResponseAccountListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AccountListResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "name": "Bearer Authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}